Issue: WSW-17509 MISRA C-2012 Rule 10.4 patch 26
[Issue]
Defects in Component midware_wifi.After analyzing using Coverity and Misra tools,
MISRA C-2012 rule 10.4,essential type of the left hand operand is not the same
as that of the right operand.

[Fix]
Changes done as to make left hand operand is same as that of the
right operand.Miscellaneous changes.

[Test]
Compile and Build successful with RT1060 & 8987 wifi_cli on IAR,Keil
& Armgcc.(0 Error(s), 0 Warning(s)).Scan and Connect as STA.
uAP start and external station connect and ping.

[Open Issues/Todo]
No

[Reviewers]
Abhishek Misra,Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

+++ wifidriver/mlan_sta_cmdresp.c
@@ -430,7 +430,7 @@ static mlan_status wlan_get_power_level(pmlan_private pmpriv, void *pdata_buf)
         {
             max_power = pg->power_max;
             min_power = pg->power_min;
-            length -= sizeof(Power_Group_t);
+            length -= (int)sizeof(Power_Group_t);
         }
         while (length > 0)
         {
@@ -443,7 +443,7 @@ static mlan_status wlan_get_power_level(pmlan_private pmpriv, void *pdata_buf)
             {
                 min_power = pg->power_min;
             }
-            length -= sizeof(Power_Group_t);
+            length -= (int)sizeof(Power_Group_t);
         }
         if (ppg_tlv->length > 0U)
         {
